FIxed readcoefs in jastroAniso_m.

+++ src/jastrowAniso_m.f90
@@ -733,7 +733,7 @@ contains
     character(len=*), intent(in) :: lines(:)
     integer, intent(inout) :: idx
     integer :: m
-    integer, allocatable :: gl_oldorder(:)
+    real(r8), allocatable :: gl_oldorder(:)
 
     this%gl = 0.0d0
     do m=1,this%gnum
